% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_name.R
\name{add_transcript_version}
\alias{add_transcript_version}
\title{Add transcript version}
\usage{
add_transcript_version(x)
}
\arguments{
\item{input}{Biocyc Gene Name ending with .1}
}
\description{
Adds 1 to a Gene.name string

Add .1 at the end of the string
}
