% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corncyc.R, R/zzz.R
\name{cyc_file}
\alias{cyc_file}
\title{Make cyc file paths from configuaration keys}
\usage{
cyc_file(x)

cyc_file(x)
}
\arguments{
\item{x}{config key}

\item{input}{Biocyc col file}
}
\description{
Make cyc file paths from configuaration keys

Search for file in configuration makes a path for it
}
\examples{
\dontrun{
cyc_file("genes_col")
}

}
