% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_name.R
\name{drop_transcript_suffix}
\alias{drop_transcript_suffix}
\title{Remove transcript name suffix}
\usage{
drop_transcript_suffix(x)
}
\arguments{
\item{input}{Biocyc Gene Name ending with .1}
}
\description{
Trim gene name so that you get only the gene model name Zm or GRM
}
