% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corncyc.R, R/read_col.R
\name{read_col}
\alias{read_col}
\title{Read Biocyc col report tables}
\usage{
read_col(input = NULL)

read_col(input = NULL)
}
\arguments{
\item{input}{Biocyc col file}
}
\value{
a dataframe. Note that many column names are serialized because
they are different values for a common key in dat files

a dataframe. Note that many column names are serialized because
they are different values for a common key in dat files
}
\description{
The col files are tab delimited files representing info in data files

The col files are tab delimited files representing info in data files
}
\examples{
\dontrun{
read_col("proteins.dat")
}

\dontrun{
read_col("proteins.dat")
}

}
