% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dat.R
\name{read_dat}
\alias{read_dat}
\title{Read Biocyc dat files}
\usage{
read_dat(input = NULL)
}
\arguments{
\item{input}{Biocyc dat file}
}
\value{
a list of named vaectors representing key:value pairs
}
\description{
The dat files are delimited by double slashes // each line being a key-value pair
}
\examples{
\dontrun{
read_dat("proteins.dat")
}
}
